requirements.txt
python-dotenv
openai
streamlit
PyMuPDF
python-docx
chromadb
sentence-transformers
unstructured[pdf]
unstructured-inference
opencv-python-headless
pdfplumber
nltk
openpyxl
